Fix h/w tests to work with TF2

+++ tensorflow/compiler/xla/tests/exhaustive_op_test_utils.h
@@ -572,7 +572,7 @@ FpValues GetNans(int approx_num_values) {
 
 template <typename T>
 FpValues GetNormals(int approx_num_values) {
-  float component_total = std::sqrtf(approx_num_values);
+  float component_total = std::sqrt(approx_num_values);
   return GetFpValues<T>(
       BitChunks(0x1, GetAllOneMantissa<T>(),
                 (1ull << (GetMantissaTotalBits<T>() + 1)) / component_total),
